Preliminary version of global->global pointer detection
We run llvm-link first to link all the bc files together, then use stride-based
field representation inclusion-based andersen's analysis from SVF on the module to get field-sensitive
alias results.

Next, we build the call-graph of the application and jump to subgraph for the given function
name, DFS on that call-graph and go through all load and store instructions that load or store-to
a global, where the value itself aliases a global.

Included tiny-web-server alias analysis as a test case.

Requirements for it to work with SVF: clang10.0 + llvm 10.0, SVF version 1.9.