Adding support for function pointer call detection
We scan for function pointer calls (indirect call instructions) and then get the load instructions preceeding the call instructions.
These load instructions are then checked if they alias any globals. We're trying to detect globals which contain function pointers
and what the offset is into those globals so the function pointers can be modified with the offset for lmvx.